feat: implement a function which returns the tablets which can send SemiSyncAcks for a given primary
